% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addinexamples-package.R
\docType{package}
\name{addinexamples}
\alias{addinexamples}
\alias{addinexamples-package}
\title{RStudio Addins}
\description{
Use RStudio addins to execute \R functions interactively within RStudio.
See \code{\link{findAndReplaceAddin}()} and \code{\link{insertInAddin}()}
for two examples of R functions that are exported.
}

