% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findAndReplaceAddin.R
\name{findAndReplaceAddin}
\alias{findAndReplaceAddin}
\title{Find and Replace Tokens in a Document}
\usage{
findAndReplaceAddin()
}
\description{
Call this as an addin to replace occurrences of a particular sequence of
characters in a document with a new sequence of characters.
}

