% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertInAddin.R, R/updatePackagesAddin.R
\name{insertInAddin}
\alias{insertInAddin}
\alias{insertInAddin}
\title{Insert \%in\%.}
\usage{
insertInAddin()

insertInAddin()
}
\description{
Call this function as an addin to insert \code{ \%in\% } at the cursor position.

When called in the repo of an R package, its package dependencies are inspected
and the obsolete ones are updated.  This function is a thin wrapper around
\code{update(remotes::dev_package_deps())}.  Unlike the 'Update' button in RStudio's 'Packages' panel,
this function will (a) update from CRAN and remote sources like GitHub and
(b) not attempt to install local packages that are unrelated to the current package.
}
