% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/natpipeInAddin.R
\name{natpipeInAddin}
\alias{natpipeInAddin}
\title{Insert |>.}
\usage{
natpipeInAddin()
}
\description{
Call this function as an addin to insert \code{ |> } (native >=R 4.1 pipe) at
the cursor position.
}
