% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refactor.R
\name{refactor}
\alias{refactor}
\title{Rename Words in a Document}
\usage{
refactor()
}
\description{
Call this as an addin to replace occurrences of a particular sequence of character
in a document with a new sequence of characters.
}

