% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudioaddins-package.R
\docType{package}
\name{rstudioaddins}
\alias{rstudioaddins}
\alias{rstudioaddins-package}
\title{RStudio Addins}
\description{
Use RStudio addins to execute \R functions interactively within RStudio.
See \code{\link{refactor}()} and \code{\link{insertIn}()} for two examples
of R functions that are exported.
}

