% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsetAddin.R
\name{subsetAddin}
\alias{subsetAddin}
\title{Subset a Data Frame.}
\usage{
subsetAddin()
}
\description{
Interactively subset a \code{data.frame}. The resulting
code will be emitted as a call to the \code{\link{subset}}
function.
}
\details{
This addin can be used to interactively subset a \code{data.frame}.
The intended way to use this is as follows:

1. Highlight a symbol naming a \code{data.frame} in your R session,
   e.g. \code{mtcars},
2. Execute this addin, to interactively subset it.

When you're done, the code performing this operation will be emitted
at the cursor position.
}

